+++ obc/drivers/CMakeLists.txt
@@ -52,6 +52,8 @@ target_sources(OBC-firmware.out PUBLIC
 
     ${CMAKE_CURRENT_SOURCE_DIR}/sw_watchdog/obc_sw_watchdog.c
 
-    ${CMAKE_CURRENT_SOURCE_DIR}/vn100/vn100.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/vn100/vn100_ascii.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/vn100/vn100_binary.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/vn100/vn100_common.c
     ${CMAKE_CURRENT_SOURCE_DIR}/vn100/vn100_parsing.c
 )
