+++ obc/examples/vn100_demo/test_binary_reading/main.c
@@ -1,7 +1,7 @@
 #include "obc_print.h"
 #include "obc_sci_io.h"
 #include "obc_board_config.h"
-#include "vn100.h"
+#include "vn100_common.h"
 
 #include <FreeRTOS.h>
 #include <os_task.h>
@@ -15,24 +15,26 @@ static StaticTask_t taskBuffer;
 static StackType_t taskStack[1024];
 
 void vTaskCode(void* pvParameters) {
-  //Move this back to main after bug with sciReadBytes & dabort_ is fixed
+  // Move this back to main after bug with sciReadBytes & dabort_ is fixed
   initVN100();
 
-  obc_error_code_t errCode;    
-  vn_ymr_packet_t myPacket;
+  obc_error_code_t errCode;
+  vn_binary_packet_t packet;
 
   sciPrintf("Beginning demo \r\n");
 
   while (1) {
-    errCode = readVN100(VN_YMR, &myPacket);
+    errCode = readVN100(VN_YMR, &packet);
 
     if (errCode != OBC_ERR_CODE_SUCCESS) {
-        sciPrintf("Error reading from VN100 - %d\r\n", errCode);
-    }
-    else {
-        sciPrintf("Yaw:   %f | MagX: %f | AccelX: %f | GyroX: %f \r\n", myPacket.yaw, myPacket.magX, myPacket.accelX, myPacket.gyroX);
-        sciPrintf("Pitch: %f | MagY: %f | AccelY: %f | GyroY: %f \r\n", myPacket.yaw, myPacket.magX, myPacket.accelX, myPacket.gyroX);
-        sciPrintf("Roll:  %f | MagZ: %f | AccelZ: %f | GyroZ: %f \r\n", myPacket.yaw, myPacket.magX, myPacket.accelX, myPacket.gyroX);
+      sciPrintf("Error reading from VN100 - %d\r\n", errCode);
+    } else {
+      sciPrintf("Yaw:   %f | MagX: %f | AccelX: %f | GyroX: %f \r\n", packet.yaw, packet.magX, packet.accelX,
+                packet.gyroX);
+      sciPrintf("Pitch: %f | MagY: %f | AccelY: %f | GyroY: %f \r\n", packet.yaw, packet.magX, packet.accelX,
+                packet.gyroX);
+      sciPrintf("Roll:  %f | MagZ: %f | AccelZ: %f | GyroZ: %f \r\n", packet.yaw, packet.magX, packet.accelX,
+                packet.gyroX);
     }
     // Toggle the LED.
     gioToggleBit(gioPORTB, 1);
